auditbeat: split system process module
This splits the common module parts and the gosysinfo backend.
Needed for an upcoming PR which adds a quark backend, so I'm splitting this
earlier to make review easier.

Only functional case for this PR is add missing `bucket.Close()` calls in New().
